+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/core/render/screens/IScreen.java
@@ -1,0 +1,10 @@
+package me.molybdenum.ambience_mini.engine.client.core.render.screens;
+
+public interface IScreen<TScreen, TWidget> {
+    TScreen getScreen();
+
+    void addWidget(TWidget widget);
+
+    int screenWidth();
+    int screenHeight();
+}