+++ v1_20_1/src/main/java/me/molybdenum/ambience_mini/v1_20_1/client/core/ClientCore.java
@@ -2,6 +2,7 @@ package me.molybdenum.ambience_mini.v1_20_1.client.core;
 
 
 import me.molybdenum.ambience_mini.engine.client.core.flags.FlagCache;
+import me.molybdenum.ambience_mini.engine.client.core.music.ServerMusicCache;
 import me.molybdenum.ambience_mini.v1_20_1.client.core.networking.ClientNetworkManager;
 import me.molybdenum.ambience_mini.v1_20_1.client.core.setup.ClientConfig;
 import me.molybdenum.ambience_mini.v1_20_1.client.core.setup.KeyBindings;
@@ -45,6 +46,7 @@ public class ClientCore extends BaseClientCore<
             ClientAreaManager areaManager,
             AreaRenderer renderer,
             FlagCache flagCache,
+            ServerMusicCache musicCache,
             ServerSetup serverSetup,
             ClientConfig clientConfig,
             KeyBindings keyBindings,
@@ -53,7 +55,7 @@ public class ClientCore extends BaseClientCore<
             ScreenState screenState,
             CombatState combatState
     ) {
-        super(McVersion.V1_20, logger, nameCache, structureCache, notification, networkManager, areaManager, renderer, flagCache, serverSetup, clientConfig, keyBindings, playerState, levelState, screenState, combatState);
+        super(McVersion.V1_20, logger, nameCache, structureCache, notification, networkManager, areaManager, renderer, flagCache, musicCache, serverSetup, clientConfig, keyBindings, playerState, levelState, screenState, combatState);
     }