+++ v1_20_1/src/main/java/me/molybdenum/ambience_mini/v1_20_1/client/core/render/drawer/Drawer.java
@@ -15,7 +15,7 @@ import org.joml.Matrix3f;
 import org.joml.Matrix4f;
 import org.lwjgl.opengl.GL11;
 
-public class Drawer extends BaseDrawer
+public class Drawer extends BaseDrawer<PoseStack.Pose>
 {
     private final Minecraft mc = Minecraft.getInstance();
 
@@ -25,6 +25,7 @@ public class Drawer extends BaseDrawer
     private MultiBufferSource.BufferSource buffer = null;
 
 
+    @Override
     public void setup(PoseStack.Pose pose) {
         this.pose = pose.pose();
         this.normal = pose.normal();