+++ changelog.md
@@ -34,10 +34,12 @@
 - Added server-side mod options for enabling and disabling use of server flags and server areas.
 
 **Miscellaneous:**
+- You now need to click the "Delete" button twice in the area-menu to actually delete the area.
 - Minor improvements to output created by verbose mode.
 - Minor improvements to volume change handling.
 - Architectural improvements to message passing and handling between client and server.
 - Architectural improvements to music-configuration loading and execution.
+- Architectural improvements to menus/screens.
 - Added support for `.wav` files.
   * Metadata is not supported, however, so you should use the new `loopstart` and similar music flags to configure looping for `.wav` files.
 - Various bug-fixes.